Build chatbot api to take text and event request from frontend and chat with dialogflow
