+++ bot-api/package.json
@@ -6,12 +6,11 @@
     "start": "node ./bin/www"
   },
   "dependencies": {
-    "@google-cloud/dialogflow": "^3.4.0",
-    "cookie-parser": "~1.4.4",
-    "debug": "~2.6.9",
-    "express": "~4.16.1",
-    "http-errors": "~1.6.3",
-    "morgan": "~1.9.1",
-    "uuid": "^8.3.2"
+    "actions-on-google": "^2.14.0",
+    "dialogflow": "^1.2.0",
+    "dialogflow-fulfillment": "^0.6.1",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "mongoose": "^5.12.7"
   }
 }
